latest code updated

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,8 @@ use Inertia\Inertia;
 use App\Models\Post;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\PostRequest;
+
 
 class PostController extends Controller
 {
@@ -17,37 +19,33 @@ class PostController extends Controller
      */
     public function index()
     {
-        // $posts = Post::all();
-        // return Inertia::render('Posts/Index')
-        //     ->with('posts' , $posts);
-
        $posts = Post::latest()->paginate(3);
        return Inertia::render('Posts/Index')
                ->with('posts' , $posts);
     }
 
-    public function search()
-    {
-        // $posts = Post::all();
-        $posts = Post::latest()->paginate(3);
-        // return Inertia::render('Posts/SearchPaginatedPage')
-        return Inertia::render('Posts/SearchIndex')
-               ->with('posts' , $posts);
-    }
+    // public function search()
+    // {
+    //     // $posts = Post::all();
+    //     $posts = Post::latest()->paginate(3);
+    //     // return Inertia::render('Posts/SearchPaginatedPage')
+    //     return Inertia::render('Posts/SearchIndex')
+    //            ->with('posts' , $posts);
+    // }
 
-    public function pagination()
-    {
-       $posts = Post::latest()->paginate(10);
-       return Inertia::render('Posts/SearchPagination')
-               ->with('posts' , $posts);
-    }
+    // public function pagination()
+    // {
+    //    $posts = Post::latest()->paginate(10);
+    //    return Inertia::render('Posts/SearchPagination')
+    //            ->with('posts' , $posts);
+    // }
 
-    public function sort()
-    {
-       $posts = Post::latest()->paginate(10);
-       return Inertia::render('Posts/SearchPaginatedBoot')
-               ->with('posts' , $posts);
-    }
+    // public function sort()
+    // {
+    //    $posts = Post::latest()->paginate(10);
+    //    return Inertia::render('Posts/SearchPaginatedBoot')
+    //            ->with('posts' , $posts);
+    // }
     
     /**
      * Write code on Method
@@ -63,7 +61,6 @@ class PostController extends Controller
     {
         return Inertia::render('Posts/Show')
             ->with('post' , $post);
-        // return Inertia::render('Posts/show');
     }
 
     /**
@@ -71,35 +68,23 @@ class PostController extends Controller
      *
      * @return Response
      */
-    public function store(Request $request)
+    
+    public function store(PostRequest $request)
     {
-
-        // var_dump($request->input()); exit;
-
-        Validator::make($request->all(), [
-            'title' => ['required'],
-            'body' => ['required'],
-            'is_display' => ['required'],
-            'is_approved'=> ['required'],
-            'language'=> ['required'],
-            'posted_date'=> ['required'],
-            'posted_time'=> ['required'],
-
-        ])->validate();
-
-        $posted_at=  $request->input('posted_date') . " ". $request->input('posted_time');
+        $posted_at=  $request->validated('posted_date') . " ". $request->validated('posted_time');
         $date = strtotime($posted_at);
 
         Post::create([
-            'language'      => $request->input('language'),
-            'title'         => $request->input('title'),
-            'body'          => $request->input('body'),
-            'is_display'    => $request->input('is_display'),
-            'is_approved'   => $request->input('is_approved'),
+            'language'      => $request->validated('language'),
+            'title'         => $request->validated('title'),
+            'body'          => $request->validated('body'),
+            'is_display'    => $request->validated('is_display'),
+            'is_approved'   => $request->validated('is_approved'),
             'posted_at'     => date('Y-m-d H:i:s', $date),
         ]);
-
+        
         return redirect()->route('posts.index');
+        // return redirect()->route('shop.index')->with("message",__("Shop created"));
     }
   
     /**
@@ -118,28 +103,17 @@ class PostController extends Controller
      *
      * @return Response
      */
-    public function update($id, Request $request)
+    public function update($id, PostRequest $request)
     {
-        Validator::make($request->all(), [
-            'title' => ['required'],
-            'body' => ['required'],
-            'is_display' => ['required'],
-            'is_approved'=> ['required'],
-            'language'=> ['required'],
-            'posted_date'=> ['required'],
-            'posted_time'=> ['required'],
-
-        ])->validate();
-
-        $posted_at=  $request->input('posted_date') . " ". $request->input('posted_time');
+        $posted_at=  $request->validated('posted_date') . " ". $request->validated('posted_time');
     
         Post::find($id)->update([
-            'language'      => $request->input('language'),
-            'title'         => $request->input('title'),
-            'body'          => $request->input('body'),
-            'is_display'    => $request->input('is_display'),
-            'is_approved'   => $request->input('is_approved'),
-            'posted_at'     => date('Y-m-d H:i:s', strtotime($posted_at)),
+            'language'      => $request->validated('language'),
+            'title'         => $request->validated('title'),
+            'body'          => $request->validated('body'),
+            'is_display'    => $request->validated('is_display'),
+            'is_approved'   => $request->validated('is_approved'),
+            'posted_at'     => date('Y-m-d H:i:s', $date),
         ]);
 
         return redirect()->route('posts.index');
@@ -150,13 +124,14 @@ class PostController extends Controller
      *
      * @return Response
      */
-    public function destroy($id)
+    public function destroy($id) //,PostRequest $request
     {
         // if (Gate::allows('isAdmin')) {
         //     Post::find($id)->delete();
         // } 
 
-        // Gate::authorize('isAdmin');
+        Gate::authorize('isAdmin');
+            
         Post::find($id)->delete();
         return redirect()->route('posts.index');
     }
